requirements.txt
cloudpickle==2.0.0
cmake==3.22.3
gym==0.23.1
gym-notices==0.0.6
importlib-metadata==4.11.3
MouseInfo==0.1.3
neat-python==0.92
numpy==1.22.3
Pillow==9.1.0
PyAutoGUI==0.9.53
pygame==2.1.2
PyGetWindow==0.0.9
pyglet==1.5.23
PyMsgBox==1.0.9
pyperclip==1.8.2
PyRect==0.2.0
PyScreeze==0.1.28
pytweening==1.0.4
scipy==1.8.0
zipp==3.8.0
